requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
google-generativeai>=0.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
langchain>=0.0.300
pydantic>=1.10.0